+++ src/main/scala/com/campudus/tableaux/exceptions.scala
@@ -3,7 +3,7 @@ package com.campudus.tableaux
 import com.campudus.tableaux.database.domain.ColumnType
 import com.campudus.tableaux.database.model.TableauxModel.{RowId, TableId}
 import com.campudus.tableaux.router.RouterException
-import com.campudus.tableaux.router.auth.{Action, Scope}
+import com.campudus.tableaux.router.auth.permission.{Action, Scope}
 
 sealed trait CustomException extends Throwable {
   val message: String
