+++ src/test/scala/com/campudus/tableaux/api/auth/StructureControllerAuthTest.scala
@@ -4,7 +4,7 @@ import com.campudus.tableaux.controller.StructureController
 import com.campudus.tableaux.database.DatabaseConnection
 import com.campudus.tableaux.database.domain.{DisplayInfos, GenericTable}
 import com.campudus.tableaux.database.model.StructureModel
-import com.campudus.tableaux.router.auth.RoleModel
+import com.campudus.tableaux.router.auth.permission.RoleModel
 import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
