+++ src/test/scala/com/campudus/tableaux/controller/StructureControllerTest.scala
@@ -3,7 +3,7 @@ package com.campudus.tableaux.controller
 import com.campudus.tableaux.database.domain.{CreateSimpleColumn, GenericTable}
 import com.campudus.tableaux.database.model.StructureModel
 import com.campudus.tableaux.database.{DatabaseConnection, LanguageNeutral, TextType}
-import com.campudus.tableaux.router.auth.RoleModel
+import com.campudus.tableaux.router.auth.permission.RoleModel
 import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
