+++ src/test/scala/com/campudus/tableaux/database/model/FileModelTest.scala
@@ -3,7 +3,7 @@ package com.campudus.tableaux.database.model
 import com.campudus.tableaux.controller.MediaController
 import com.campudus.tableaux.database.DatabaseConnection
 import com.campudus.tableaux.database.domain.MultiLanguageValue
-import com.campudus.tableaux.router.auth.RoleModel
+import com.campudus.tableaux.router.auth.permission.RoleModel
 import com.campudus.tableaux.testtools.TableauxTestBase
 import io.vertx.ext.unit.TestContext
 import io.vertx.ext.unit.junit.VertxUnitRunner
